Fix/syncjob count querydsl

+++ src/main/java/com/findex/demo/global/pagination/dto/PagedResponse.java
@@ -1,6 +1,6 @@
 package com.findex.demo.global.pagination.dto;
 import java.util.List;
 
-public record PagedResponse<T>(List<T> content, String nextCursor, String nextIdAfter, int size, int totalElements, Boolean hasNext) {
+public record PagedResponse<T>(List<T> content, String nextCursor, String nextIdAfter, int size, Long totalElements, Boolean hasNext) {
 
 }
